pjt02/requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
pep8==1.7.1
python-decouple==3.1
requests==2.22.0
urllib3==1.25.3
